refactor: extract Array2D into standalone crate for shared use
Move Array2D and ArrayElement from timsquery into a new rust/array2d
workspace crate with its own Array2DError type. timsquery re-exports
from array2d and bridges errors via From<Array2DError> for
DataProcessingError. calibrt gains array2d as a direct dependency.
