+++ rust/array2d/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "array2d"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[dependencies]
+serde = { workspace = true }
