+++ rust/timsquery/src/serde/chromatogram_output.rs
@@ -113,9 +113,9 @@ impl ChromatogramOutput {
                     return None;
                 }
                 let out_vec = out_slc.to_vec();
-                Some(Ok(((mz, out_vec), format!("{}", idx))))
+                Some(Ok::<_, crate::errors::DataProcessingError>(((mz, out_vec), format!("{}", idx))))
             })
-            .collect::<Result<Vec<_>, _>>()?
+            .collect::<Result<Vec<_>, crate::errors::DataProcessingError>>()?
             .into_iter()
             .unzip();
